Small cent pages (#8)
* Flying Eagle page

Made the flying eagle page and changed the template a little.

* Made the Indian Head entry

 Made the Entry for the Indian Head cent and fixed the main Header and the title for the Flying Eagle cent.

* Made some pages

Made the Wheat penny page. Made edits to the Flying eagle and Indian head penny page.

* Update Flying Eagal.html

* Update IndianHeadCent.html

* Update Wheat penny.html

* Made the wheat penny page

Made the page for the Wheat penny

* Added memorial page

Added the page for the memorial cent.

* Spelling

* Made the 2009 penny page

Made the page for the 2009 pennies

* Made the shield penny page

Made the page for shield pennies.